+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Sith4",
+  "name": "sith",
   "version": "4.0.0",
   "scripts": {
     "ng": "ng",
