+++ src/app/app.module.ts
@@ -6,10 +6,6 @@ import { AppComponent } from './app.component';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
-import { ActivatedRoute } from '@angular/router';
-import { CommonModule } from '@angular/common';
-import { IconsModule } from './icons/icons.module';
-import { PagesModule } from './pages/pages.module';
 import { ComponentsModule } from './components/components.module';
 
 export function HttpLoaderFactory(http: HttpClient) {
@@ -21,10 +17,7 @@ export function HttpLoaderFactory(http: HttpClient) {
 	imports: [
 		AppRoutingModule,
 		BrowserModule,
-		CommonModule,
 		ComponentsModule,
-		IconsModule,
-		PagesModule,
 		HttpClientModule,
 		TranslateModule.forRoot({
 			defaultLanguage: 'en-US',
@@ -38,7 +31,4 @@ export function HttpLoaderFactory(http: HttpClient) {
 	providers: [],
 	bootstrap: [AppComponent],
 })
-export class AppModule {
-	lang = this.route.snapshot.params['lang'];
-	constructor(private route: ActivatedRoute) {}
-}
+export class AppModule {}
