+++ src/app/components/common/icons/expand_less/expand_less.ts
