+++ src/app/components/common/icons/expand_more/expand_more.ts
