+++ src/app/components/common/icons/icons.scss
