+++ src/app/components/mobile/subtitle/subtitle.component.ts
@@ -1,11 +1,11 @@
 import { Component, Input } from '@angular/core';
 
 @Component({
-	selector: 'app-subtitle',
+	selector: 'app-mobile-subtitle',
 	templateUrl: './subtitle.html',
 	styleUrls: ['./subtitle.scss'],
 })
-export class SubtitleComponent {
+export class MobileSubtitleComponent {
 	@Input() text = '';
 	@Input() bordered = false;
 	@Input() onDarkBackground = false;
