+++ src/app/components/pages/home/home.html
@@ -1,0 +1,1 @@
+<app-mobile-lang-selector />
