--- src/app/pages/home/home.html
@@ -1,3 +1,0 @@
-<div>
-  <app-mobile-lang-selector />
-</div>
