--- src/assets/i18n/fr_FR.json
@@ -1,7 +1,0 @@
-{
-  "i18n": {
-    "switch_lang": "Choix de la langue",
-    "select_lang": "Sélectionnez une langue",
-    "help_us_translate": "Aidez-nous à traduire ce site"
-  }
-}
